// src/utils/enums/requisitionStatus.ts
export enum requisitionStatusEnum {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  INSPECTING = 'INSPECTING',
  RECEIVED = 'RECEIVED',
  DISPUTED = 'DISPUTED',
  RESOLVED = 'RESOLVED',
  FULFILLED = 'FULFILLED',
}

export enum headApprovalEnum {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  DECLINED = 'DECLINED',
}

export enum disputeAccountRequsitionEnum {
  PENDING = 'PENDING',
  RESOLVED = 'RESOLVED',
  REJECTED = 'REJECTED',
}